strset

+++ libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: sunmin <msh4287@naver.com>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/03 16:35:28 by sunmin            #+#    #+#             */
-/*   Updated: 2020/10/05 20:56:33 by sunmin           ###   ########.fr       */
+/*   Updated: 2020/10/05 21:02:18 by sunmin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 size_t		ft_strlcat(char *dst, const char *src, size_t size)
 {
-	size_t		size_dst;
-	size_t		size_src;
-	int		i;
-	int		j;
+	size_t			size_dst;
+	size_t			size_src;
+	unsigned long	i;
+	unsigned long	j;
 
 	size_dst = ft_strlen(dst);
 	size_src = ft_strlen(src);
